fix: correcting functions of return

See merge request guilhermeciole/softclient!68

+++ app/controllers/orders_controllers.py
@@ -1,18 +1,17 @@
-from flask import jsonify, request, current_app, send_file
+from flask import json, jsonify, request, current_app, send_file
+from sqlalchemy.sql.functions import user
 from app.models.order_model import OrderModel
 from http import HTTPStatus
 from werkzeug.exceptions import NotFound
-from app.utils.format_date import format_date_and_time
 from app.utils.permission import permission_role
 from app.exceptions.orders_exceptions import KeyTypeError, InvalidDate
 import sqlalchemy
-from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
+from app.utils.format_date import format_datetime
+from app.utils.format_date import format_date_and_time
 from reportlab.pdfgen import canvas
 from reportlab.lib.pagesizes import A4
 from datetime import datetime
-
-
-
 
 
 @jwt_required()
@@ -23,22 +22,30 @@ def list_orders():
         "type": order.type.value,
         "status": order.status.value,
         "description": order.description,
-        "release_date": format_date_and_time(order.release_date),
-        "update_date": format_date_and_time(order.update_date),
+        "release_date": order.release_date,
+        "update_date": order.update_date,
         "solution": order.solution,
-        "user_id": order.user_id,
-        "technician_id": order.technician_id
+            "user": {
+                "id": order.user.id,
+                "name": order.user.name,
+                "email": order.user.email,
+                "position": order.user.position,
+                "role": order.user.role.value,
+            },
+            "technician": order.technician,
     } for order in orders_list]), HTTPStatus.OK
 
 
-@permission_role(('user',))
+@permission_role(('user', 'admin'))
 @jwt_required()
 def create_order():
-    user = get_jwt_identity()
+    user_logged = get_jwt_identity()
 
+    if 'position' not in user_logged.keys():
+        return jsonify({"message": "only users to place an order"}), HTTPStatus.UNAUTHORIZED
     try:
         data = request.json
-        data['user_id'] = user['id']
+        data['user_id'] = user_logged['id']
         OrderModel.validate(data)
         verified_data = OrderModel.check_needed_keys(data)
         new_data = OrderModel.create_order_data(verified_data)
@@ -48,14 +55,20 @@ def create_order():
 
         return jsonify({
         "id": order.id,
-        "type": order.type.value,
+        "type": order.type.value.title(),
         "status": order.status.value,
         "description": order.description,
-        "release_date": format_date_and_time(order.release_date),
-        "update_date": format_date_and_time(order.update_date),
+        "release_date": order.release_date,
+        "update_date": order.update_date,
         "solution": order.solution,
-        "user_id": order.user.id,
-        "technician_id": order.technician_id,
+        "user": {
+            "id": order.user.id,
+            "name": order.user.name,
+            "email": order.user.email,
+            "position": order.user.position,
+            "role": order.user.role.value,
+        },
+        "technician": order.technician,
     }), HTTPStatus.OK
     except InvalidDate as e:
         return jsonify({"message": str(e)}), HTTPStatus.BAD_REQUEST
@@ -71,22 +84,31 @@ def get_order_by_id(id: int):
         order = OrderModel.query.get_or_404(id)
         return jsonify({
             "id": order.id,
-            "type": order.type.value,
+            "type": order.type.value.title(),
             "status": order.status.value,
             "description": order.description,
-            "release_date": format_date_and_time(order.release_date),
-            "update_date": format_date_and_time(order.update_date),
+            "release_date": order.release_date,
+            "update_date": order.update_date,
             "solution": order.solution,
-            "user_id": order.user.id,
-            "technician_id": order.technician_id,
+            "user": {
+                "id": order.user.id,
+                "name": order.user.name,
+                "email": order.user.email,
+                "position": order.user.position,
+                "role": order.user.role.value,
+            },
+            "technician": order.technician,
                 }), HTTPStatus.OK
     except NotFound:
         return {"Error": "Order not found."}, HTTPStatus.NOT_FOUND
     
 
-@permission_role(('user', 'tech'))
+@permission_role(('user', 'admin'))
 @jwt_required()    
 def update_order(id: int):
+    user_logged = get_jwt_identity()
+    if 'position' not in user_logged.keys():
+        return jsonify({"message": "only users to place an order"}), HTTPStatus.UNAUTHORIZED
     try:
         data = request.get_json()
         order = OrderModel.query.filter_by(id=id).first()
@@ -106,19 +128,22 @@ def update_order(id: int):
     except sqlalchemy.exc.DataError:
         return jsonify({"msg": "type value is wrong"}), 400
     
-        
-    
-
     return jsonify({
         "id": order.id,
-        "type": order.type.value,
+        "type": order.type.value.title(),
         "status": order.status.value,
         "description": order.description,
         "release_date": order.release_date,
         "update_date": order.update_date,
         "solution": order.solution,
-        "user_id": order.user.id,
-        "technician_id": order.technician_id,
+        "user": {
+            "id": order.user.id,
+            "name": order.user.name,
+            "email": order.user.email,
+            "position": order.user.position,
+            "role": order.user.role.value,
+        },
+        "technician": order.technician,
         }), 200
 
 
@@ -127,30 +152,44 @@ def get_order_by_status(order_status: str):
     try:
         orders= OrderModel.query.filter_by(status=order_status).all()
        
-        
         return jsonify([{
         "id": order.id,
-        "type": order.type.value,
+        "type": order.type.value.title(),
         "status": order.status.value,
         "description": order.description,
-        "release_date": format_date_and_time(order.release_date),
-        "update_date": format_date_and_time(order.update_date),
+        "release_date": order.release_date,
+        "update_date": order.update_date,
         "solution": order.solution,
-        "user_id": order.user_id,
-        "technician_id": order.technician_id
+        "user": {
+            "id": order.user.id,
+            "name": order.user.name,
+            "email": order.user.email,
+            "position": order.user.position,
+            "role": order.user.role.value,
+        },
+        "technician": order.technician,
             } for order in orders]), HTTPStatus.OK
     except NotFound:
         return {"Error": "Not found."}, HTTPStatus.NOT_FOUND
 
 
-@permission_role(('user',))
+@permission_role(('user', 'admin'))
 @jwt_required()    
 def delete_order(id: int):
-    order = OrderModel.query.get_or_404(id)
-    current_app.db.session.delete(order)
-    current_app.db.session.commit()
-    return "", HTTPStatus.OK
-    
+    user_logged = get_jwt_identity()
+    try:
+        order = OrderModel.query.get_or_404(id)
+        if order.technician:
+            return jsonify({"message": "order already assigned cannot be deleted"}), HTTPStatus.UNAUTHORIZED
+        if 'position' not in user_logged.keys():
+            return jsonify({"message": "only users to place an order"}), HTTPStatus.UNAUTHORIZED
+        if user_logged['id'] != order.user_id:
+            return jsonify({"message": "unauthorized delete order"}), HTTPStatus.UNAUTHORIZED
+        current_app.db.session.delete(order)
+        current_app.db.session.commit()
+        return "", HTTPStatus.NO_CONTENT
+    except NotFound:
+        return {"message": "Not found."}, HTTPStatus.NOT_FOUND
 
 @jwt_required()
 def get_user_by_order_id(order_id):
@@ -161,9 +200,9 @@ def get_user_by_order_id(order_id):
                 "id": order.user.id,
                 "name": order.user.name,
                 "email": order.user.email,
-                "birthdate": format_date_and_time(order.user.birthdate),
-                "registration": order.user.registration,
-                "role": order.user.role
+                "birthdate": format_datetime(order.user.birthdate),
+                "position": order.user.position,
+                "role": order.user.role.value
             }
             
         }), HTTPStatus.OK
@@ -175,20 +214,19 @@ def get_user_by_order_id(order_id):
 def get_technician_by_order_id(order_id):
     try:
         order = OrderModel.query.filter_by(id=order_id).first_or_404()
-        return jsonify({
-            "technician": {
-                "id": order.technician.id,
-                "name": order.technician.name,
-                "email": order.technician.email,
-                "registration": order.user.registration,
-                "birthdate": format_date_and_time(order.user.birthdate),
-            }
-        }), HTTPStatus.OK
+        if order.technician_id:
+            return jsonify({
+                "technician": {
+                    "id": order.technician.id,
+                    "name": order.technician.name,
+                    "email": order.technician.email,
+                    "birthdate": format_datetime(order.user.birthdate),
+                }
+            }), HTTPStatus.OK
+        return jsonify({"message": "order was not assigned to a technician"}), HTTPStatus.NOT_FOUND
     except NotFound:
-        return {"message": "Technician not found!"}, HTTPStatus.NOT_FOUND
+        return {"message": "Order not found!"}, HTTPStatus.NOT_FOUND
 
-
-    
 
 @permission_role(('super',))
 @jwt_required()
